pyproject.toml
[project]
name = "garant-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "paramiko>=4.0.0",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
